add more fixes
