+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import { Hero } from "@/components/sections/hero";
 import JoinUs from "@/components/layouts/footer";
 import { Services } from "@/components/sections/services";
 import Team from "@/components/sections/team-info";
-import { Connect } from "@/components/sections/connect";
+import { Achievement } from "@/components/sections/achievement";
 import Testimonials from "@/components/sections/testimonials";
 import { About } from "@/components/sections/about";
 
@@ -13,7 +13,7 @@ export default function HomePage() {
       <Hero />
       <About />
       <Services />
-      <Connect />
+      <Achievement />
       <CTA />
       <Team />
       {/* <Testimonials /> */}
